feat: move MockPasswordVerifier to internal/mocks
Implements task T025 from TODO.md:
- Created centralized MockPasswordVerifier implementation in internal/mocks/password_verifier.go
- Enhanced the implementation with function field (CompareFn) for customizable behavior
- Added tracking fields (CompareCalledWith, CompareCallCount) for easier test verification
- Maintained backward compatibility with existing tests through ShouldSucceed flag
- Updated test files to use the centralized mock implementation
- Cleaned up unused imports
